modularizing testing framework
