+++ .env-example
@@ -1,5 +1,6 @@
 AUTOGEN_MODEL_NAME=PUT YOUR MODEL NAME HERE, for example: gpt-4-turbo-preview or llama3-70b-8192
 AUTOGEN_MODEL_API_KEY=PUT YOUR MODEL API KEY HERE
 AUTOGEN_MODEL_BASE_URL=IF YOU ARE USING OpenAI, remove this line, otherwise add the base URL, for example: https://api.groq.com/openai/v1
+OPENAI_API_KEY=If you are using the testing/benchmarking, put your openai key here otherwise remove this line
 
 BROWSER_STORAGE_DIR=/Users/macuser/Library/Application Support/Google/Chrome/Profile 5
